changed pseudocounts in popsize updates to be proportional to probability of coalescence/not coalescence given population size of 10000

+++ src/est_popsize.h
@@ -17,12 +17,12 @@ void resample_popsizes(ArgModel *model, const LocalTrees *trees,
      };*/
 
  struct popsize_mle_data {
-     int *arr_alloc;
-     int arr_size;
-     int ***coal_counts;
-     int ***nocoal_counts;
-     int *coal_totals;
-     int *nocoal_totals;
+     double *arr_alloc;
+     double arr_size;
+     double ***coal_counts;
+     double ***nocoal_counts;
+     double *coal_totals;
+     double *nocoal_totals;
      int ntimes;
      int popsize_idx;
      int numleaf;
